finished driver_receiving_order_page.dart + added refresh button and mark as completed order
